+++ packages/web-components/src/components/card/card.ts
@@ -11,6 +11,7 @@ import { TemplateResult, html, LitElement } from 'lit';
 import { property, query, state } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import ArrowRight20 from '@carbon/web-components/es/icons/arrow--right/20.js';
+import ArrowLeft20 from '@carbon/web-components/es/icons/arrow--left/20.js';
 import CDSLink from '@carbon/web-components/es/components/link/link.js';
 import markdownToHtml from '@carbon/ibmdotcom-utilities/es/utilities/markdownToHtml/markdownToHtml.js';
 import { carbonElement as customElement } from '@carbon/web-components/es/globals/decorators/carbon-element.js';
@@ -28,11 +29,13 @@ import CTAMixin from '../../component-mixins/cta/cta';
 const { prefix, stablePrefix: c4dPrefix } = settings;
 
 /**
- * The table mapping slot name with the private property name that indicates the existence of the slot content.
+ * The table mapping slot name with the private property name that indicates the
+ * existence of the slot content.
  */
 const slotExistencePropertyNames = {
   image: '_hasImage',
   pictogram: '_hasPictogram',
+  eyebrow: '_hasEyebrow',
 };
 
 /**
@@ -78,17 +81,32 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
   @state()
   protected _hasPictogram = false;
 
+  /**
+   * `true` if there is eyebrow content.
+   */
+  @state()
+  protected _hasEyebrow = false;
+
   @property({ attribute: 'no-poster', type: Boolean })
   noPoster;
 
   /**
-   * Handles `slotchange` event.
+   * Handles `slotchange` event for slots other than the copy slot.
    */
   protected _handleSlotChange({ target }: Event) {
     const { name } = target as HTMLSlotElement;
-    const hasContent = Boolean(this.querySelector('p'));
-    this[slotExistencePropertyNames[name]] = hasContent;
-    this._hasCopy = hasContent;
+    this[slotExistencePropertyNames[name]] = (target as HTMLSlotElement)
+      .assignedNodes()
+      .some(
+        (node) => node.nodeType !== Node.TEXT_NODE || node!.textContent!.trim()
+      );
+  }
+
+  /**
+   * Handles `slotchange` event for the copy slot.
+   */
+  protected _handleCopySlotChange() {
+    this._hasCopy = Boolean(this.querySelector('p'));
   }
 
   /**
@@ -133,7 +151,7 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
     const { _hasCopy: hasCopy } = this;
     return html`
       <div ?hidden="${!hasCopy}" class="${prefix}--card__copy" part="copy">
-        <slot @slotchange="${this._handleSlotChange}"></slot>
+        <slot @slotchange="${this._handleCopySlotChange}"></slot>
       </div>
     `;
   }
@@ -159,7 +177,7 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
             </c4d-image>
           `;
     return html`
-      <div part="image-wrapper">
+      <div part="image-wrapper" class="${prefix}--card__image-wrapper">
         <slot name="image" @slotchange="${this._handleSlotChange}">
           ${image}
         </slot>
@@ -180,9 +198,14 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
   }
 
   protected _renderPictogramSlot(placement: PICTOGRAM_PLACEMENT) {
-    const { _handleSlotChange: handleSlotChange } = this;
+    const { _handleSlotChange: handleSlotChange, _hasPictogram: hasPictogram } =
+      this;
     return html`
-      <div part="pictogram-wrapper" class="${prefix}--card__pictogram-wrapper">
+      <div
+        part="pictogram-wrapper"
+        class="${prefix}--card__pictogram-wrapper ${!hasPictogram
+          ? `${prefix}--card__pictogram-wrapper--empty`
+          : ''}">
         <slot
           name="pictogram"
           data-pictogram-placement="${placement}"
@@ -192,8 +215,13 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
   }
 
   protected _renderEyebrowSlot() {
-    return html`<div part="eyebrow-wrapper">
-      <slot name="eyebrow"></slot>
+    const { _hasEyebrow: hasEyebrow } = this;
+    return html`<div
+      part="eyebrow-wrapper"
+      class="${prefix}--card__eyebrow-wrapper ${!hasEyebrow
+        ? `${prefix}--card__eyebrow-wrapper--empty`
+        : ''}">
+      <slot name="eyebrow" @slotchange="${this._handleSlotChange}"></slot>
     </div>`;
   }
 
@@ -238,6 +266,24 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
     `;
   }
 
+  /**
+   * @returns The CTA arrow.
+   *
+   */
+  protected _renderArrow() {
+    const isLTR =
+      window.getComputedStyle(this).direction.toUpperCase() === 'LTR';
+    return html`
+      <a
+        class="${`${prefix}--card__link`}"
+        part="link"
+        href="${ifDefined(this.href)}"
+        aria-label="${this.querySelector(`${c4dPrefix}-card-heading`)
+          ?.textContent || ''}"
+        >${isLTR ? ArrowRight20() : ArrowLeft20()}</a
+      >
+    `;
+  }
   /**
    * The color scheme.
    * A typical use case of using another color scheme of card is having a "CTA" purpose of card at the last in card group.
@@ -425,19 +471,12 @@ class C4DCard extends CTAMixin(StableSelectorMixin(CDSLink)) {
       this._handleVideoTitleUpdate
     );
   }
+
   render() {
     return this._hasPictogram
       ? html`
           <div part="container">
-            ${this._renderInner()}
-            <a
-              class="${`${prefix}--card__link`}"
-              part="link"
-              href="${ifDefined(this.href)}"
-              aria-label="${this.querySelector(`${c4dPrefix}-card-heading`)
-                ?.textContent || ''}"
-              >${ArrowRight20()}</a
-            >
+            ${this._renderInner()} ${this._renderArrow()}
           </div>
         `
       : html` <div part="container">${this._renderInner()}</div> `;